Fix does not update progress bar with .mkv videos, #5254 (#5256)
This commit will change PlayerCore.refreshSyncUITimer to call syncUITime
if the timer was not running even if the timer is not being restarted. This is
needed to update the playback time when showing the OSD or OSC when
playback is paused.

This regression also reintroduced issue #3331, "Progress bar does not
reach end when i finish playing a video".